[FAB-1738] Event registration requiring signature
Use SignedEvent instead of Event for registration requests in
order to apply access control based on Creator identity against
target peer's local MSP. Note that at this point event
stream is still channel-agnostic so local MSP is checked.

Strictly speaking the registered listener is supposed to be
checked against the "Readers" policy of all channels, since
the registration will result in the listener receiving blocks
from all channels. However this is determined to be out of scope
for v1.0 due to the following reasons:
- special Readers policies that exclude normal "members" of an
  organization is considered edge cases
- rather complicated code refactoring would be needed to implement
  policy checking against channels in order to avoid a circular
  dependency from peer -> committer -> event producer -> peer

Change-Id: Ic941566c5299b4600b291137bc4724a02b5263d7
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>