[FAB-3241] Gossip: Do not send redundant StateInfo
In gossip, whenever the channel related metadata is updated,
a "dirty flag" is turned on, and periodically a goroutine disseminates
to peers if the dirty flag is turned on.
There is a bug that this dirty flag is never turned off.
This causes a redundant dissemination of StateInfo.

The stateInfo is periodically pulled from peers via a pull mechanism anyway
so there is no point in having periodical dissemination and thus the dirty
flag should be turned off after dissemination.

Change-Id: Ie508df33c28e007e952195faed7837db5641e41a
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>