[FAB-3155] LSCC security checks at validation time
This change set checks that - upon cc deploy and upgrade - the tx complies
with the instantiation policy specified in the cc packge (for deploy) and
in the cc info in the ledger (for upgrade). The committer code is
essentially performing again the checks that lscc performs at proposal time,
to ensure that no one can circumvent instantiation policies.

Furthermore, we validate
the read-write set of a cc deploy/upgrade and make sure that it is
consistent with the invocation arguments.

Finally, we make sure that LSCC deploy/upgrade transactions do not
contain writes to namespaces other than LSCC's and the chaincode that is
being deployed/upgraded.

Change-Id: Id11fc359a8c77fa58c8a966a7c324075944ae22b
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>