[FAB-3292] JoinChannel fails on e2e
The e2e example is failing in the current master
because who signs the JoinChannel is not
recognized as administrator of the peer.

This change-set does the following:
1. It fixes the error, by setting the peer's identity
as its own administrator. This is done by copying
the peer's certificate into the admincerts folder.

This change-set fixes the bug reported
at https://jira.hyperledger.org/browse/FAB-3292

Change-Id: If250c0b3b61ee56fd0cfb0da8524dd54d840c639
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>