[FAB-3317] Update peer CouchDB timeout
Occasionally CouchDB doesn't respond for (exactly) 30s.
This is most often seen in create database calls
which happens frequently during unit tests.

The default peer behavior was to retry couch after 20s,
and therefore these occurrences disrupt unit tests and
causes them to fail occasionally.

This change will set the peer CouchDB timeout to 35s for now,
in order to avoid the retries in the peer avoid the unit test problems.

After the root problem in CouchDB is resolved, the default may
change back to 20s.

Change-Id: Id688e2e137b8354512b6543fc63ad3e59a1040eb
Signed-off-by: denyeart <enyeart@us.ibm.com>