[FAB-3348] chaincode should log build level
Chaincode images may become stale and may not match fabric
runtime images. Cleaning of chaincode images will take care
of this but it will help trouble-shooting to know the build
level of the chaincode and if it matches the fabric build
level

Version information should be passed to the chaincode at the
time of building it so the chaincode can log it.

This results log :

2017-04-23 11:03:13.281 UTC [shim] SetupChaincodeLogging ->
   INFO 001 Chaincode (build level-1.0.0-snapshot-88ac3ba)
   starting up ...

Change-Id: I44dfb527ff0cb9039c80855c3e52ca3e9ba3e304
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>