[FAB-3431] Always display error callstack in logs
This CR removes the dependency on core.yaml for setting the display
of the callstack, resulting in all calls to Error() appending the
callstack (if available). This ensures the logs will always contain
the callstack. It also modifies the Message() function to only display
the error message (without callstack) for cases where just the message
message is desired. Finally, it refactors the unit tests to avoid the
previously large amount of duplicated code between test cases.

Change-Id: I379a3258fd2b357f2b5504896ff8f2ad4d5e7788
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>