[FAB-3497] Removing the hash from gossip dataMsg
The hash field isn't used anywhere but in the protos comparator function
to compare 2 ledger blocks disseminated.
This imposes a security vulnerability:
A malicious peer can send blocks with an arbitrary hash in the hash field
of the DataMsg, and then these messages would enter the in-memory stores
and fill up the memory.

Since we have no use of the block hash of our own message, I am removing it
entirely.

Change-Id: Ic22ed3d06f102795c8f2a74b27063848affc926a
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>