[FAB-3493] Fix LAST_CONFIG value on new channels
The recent consortium-related stack of changes touched on the channel
creation path and introduced a bug in the LAST_CONFIG value that is set
on all on blocks between the genesis and the first actual config block.
Specifically, their LAST_CONFIG value is set to 1 when it should be set
to 0 (i.e. point to the genesis block).

This changeset:
1. Extends a test so that this check is performed
2. Fixes this bug in a minimally invasive manner, and modifies all
dependent tests accordingly.

Details behind the bug, as well as the thinking behind this fix can be
found in the JIRA issue.

Change-Id: Ief105744503df21d289b1e2113c9038998bdda92
Signed-off-by: Kostas Christidis <kostas@christidis.io>