[FAB-3443] Improve clilogging output/usage description
This CR updates the log message for peer CLI calls to 'setlevel' to
clarify that it is using a regular expression to set the level for
multiple modules. It also updates the usage for both 'setlevel' and
'getlevel' commands. Hopefully this will help avoid confusion when
using 'getlevel', which does not currently use regular expressions
(but can be updated to do so in the future).

Change-Id: I30e7c8351dac3ed32fc19859f1f1981ccd9dc04c
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>