[FAB-3643] respect peer.gossip.endpoint configuration
This patch re-adds processing that optionally assigns
peer.gossip.endpoint to our internal endpoint representation.

Without this, enabling TLS is difficult because we cannot
control the endpoint representation at runtime in a deterministic
manner.  The end result is that x509 CN/SAN validation is
likely to fail and thus our endpoint is unreachable.

Fixes FAB-3643

Change-Id: Idd52bf55d86f87160ba508329ce379d1f3be5d04
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>