[FAB-3687] rotate /example/cluster logs
The current code configures the peer to be quite chatty
and we do not place any limits on logfile size.  Therefore,
it is quite easy to fill up disk space.  This patch fixes
this by both reducing the verbosity of the peers as well
as placing size/rotation limits in the docker configuration.

Fixes FAB-3687

Change-Id: I7bf41c240f4aed0f46014df7be2f8e4ebcb600a9
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>