[FAB-3701] Peer and orderer CAs should be separate
Prior to this change, there was a TODO in the code
to separate the trusted roots that the peer should
use when communicating with peers and orderers.
Currently they are the same (both internal maps
include all CAs for both orderer and peer/app
orgs). With this change, it's no longer the case
as the code now properly handles this.

- renamed test to match what's actually being
tested
- revamped test scenarios to only test the
functionality as well as increase coverage
- regenerated crypto material required
- slight modification to cert generator to
add SKI to intermediate root certs as it
is required
- made sure to cleanup filesystem resources

Change-Id: I20f35dcec9c787c9052fb5d8ef119cd3dbab2d6c
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>