[FAB-3727] Orderer restart broken
There was a programming error introduced during the consortium
refactoring which caused an interface return value to be checked for
nil, which returns false even when the underlying value is nil because
the dynamic type is a pointer kind.

This error caused the orderer to recognize all channels as the system
channel, and would cause a fatal error at startup.

This changeset modifies the config interface to return a boolean along
with the interface value, indicating whether the interface value was
populated.

Change-Id: Ia9bb89b765c7e8f35742895d2ac1dd2396d03908
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>