[FAB-3493] Fix LAST_CONFIG on new channels
The fix submitted in [1] works in that it delivers what we were looking
for (a pointer to the last config block) but is based on an incorrect
assumption: that the sequence number of the genesis block of a
non-system channel should be zero.

I assumed that should be the case, what with it being the genesis block
and all, but this breaks the abstraction that the config update
framework introduces. This changeset introduces the right fix. Thanks to
Jason Yellick for pointing this out.

[1] https://gerrit.hyperledger.org/r/#/c/8825/

Change-Id: I61ef2825c4cfe55495a4d81cff759fd6c340c191
Signed-off-by: Kostas Christidis <kostas@christidis.io>