[FAB-3386] Remove fabric-testenv-orderer image
We need to make our examples/scripts work with production-level
artifacts. The fabric-testenv-orderer image is not an artifact targeted
for production. This abundance of options leads to user confusion
(I've had folks ask me whether they need to use `fabric-testenv-orderer`
to make Kafka work), and it also means we have yet another image we need
to maintain.

This changeset removes all references to the fabric-testenv-orderer
Docker image.

Change-Id: I19daa1a21b2ce15b299caa77ca1e59c7101cb553
Signed-off-by: Kostas Christidis <kostas@christidis.io>