[FAB-3560] Endorsement failure doesn't return response
This CR ensures that the response is returned when endorsement fails
due to the chaincode returning an ERROR status (currently, status>=500).
The details in the response, including the status code, message, and
payload from the chaincode are vital pieces of information for
understanding the endorsement failure and acting on it, if necessary.

This CR also adds logging messages that were helpful in debugging the
issue and cleans up a few other logging messages.

Change-Id: I7189c1c5c32e0dd6cd128ec122eefa1841c74ab3
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>