[FAB-3118] pull ccenv image at runtime
The current code requires the CORE_CHAINCODE_BUILDER image to
be pre-pulled down from dockerhub or chaincode building will
fail.  This patch adds the docker-client intelligence to
check whether the image is present.  If it finds that it is
not present, it will try to pull the image from dockerhub.

Change-Id: I95b6bcd0f1b1beb2edee7f0b542f90f5fb9eabc7
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>