[FAB-3948]WIP-chaincode shim unit tests and framework
Adds a mock framework based on channel streams for chaincode
to peer connecvity. The unit test then follows the entire
shim path driven by an actual chaincode. Init and Invoke are
sent from the UT and chaincode sends ChaincoeMessages on the
mock stream. The mock stream is primed with responses for each
message received.

With this setup, with just basic GET/PUT/DEL, we get a coverage
jump from 17 to 43% as reported by
    go test -coverprofile=coverage.out

Fixed Copyright year

More tests

Change-Id: I3ea9ac94abc7f43ee5dd297b8202a360d6800cbf
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>