[FAB-3995] /examples/cluster: Fix TLS support
Chaincode currently fails to instantiate when TLS is enabled.
Two root causes were identified:

1) We were setting CORE_PEER_ADDDRESSAUTODETECT=true, which was
forcing the peer to advertise an IP rather than a DNS name to the
chaincode instance.  Since only the DNS names are registered in
the x509 for the peer, the TLS layer was rejecting the peer's cert.
We fixed this by disabling ADDRESSAUTODETECT

2) Enabling DNS revealed that the chaincode was executing on a
different docker network from the peer.  The biggest implication
was that DNS resolution for things like "peer1" was not possible.
Thus, fixing (1) broke us in a different way.  We fixed this by
setting the environment properly such that the peer launches
chaincode on the proper network.

Fixes FAB-3995

Change-Id: I1ebfbf43b63b2d288aaee7c098ebdb1f8bd5c352
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>