[FAB-4002] Upgrade to latest version of grpc-go
There have been many updates to the grpc-go package
since it was vendored in v0.6 and then v1.0.0.

There are several defects / issues related to
inactivity timeouts, connectivity, etc which can
be more easily fixed using new features such as
the client keepalive.

This CR updates grpc-go as well as all of its
dependencies and makes minor updates to the
code where required as well as recompiles all of
the protos using the protobuf compiler plugin
required by this version of grpc-go.

Note that for the updated vendoring, I chose
to delete the old packages and add the
updated ones as govendor is not perfect
when it comes to trying to update packages.

The only "new" code was adding custom server
transport credentials in the comm package as
the default provided by the grpc/credentials
package was changed and breaks the functionality
required for dynamically updating root certs for
running grpc servers and new package config
values for send/receive msg sizes as grpc-go
now enforces them.  Changes were also made
to any grpc clients and servers used in the
code to leverage the new config values.

There is one TODO which is to actually set
the max send/receive msg sizes based on
config.  This will need to be done for
both the peer and the orderer.

Also changed the license header to the new
format for any new or modified files

Change-Id: I41687a54aa21802203b04a339952881349042b70
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>