[FAB-4100] Create proto translator framework
Doing bidirectional proto translation to/from deep JSON representation
requires a significant amount of dynamic programming. Having each
message try to to handle this work independently is a recipe for
disaster, so the bulk of the reflection work needs to be centralized so
that protos may be simply extended with methods which drive the proto
translation framework.

This CR creates the base framework for handling the deep
Marshaling/Unmarshaling of all types of proto messages.  For
reviewability, it does not include any type implementations, so look for
the implementations in subsequent CRs.

The package is named protolator as a portmanteau of protobuf and
translator.  Once the component pieces are checked in and the fabric
protos annotated with the necessary helper methods, this framework will
be capable of printing arbitrary fabric protos in a human readable way
(ie Blocks, Envelopes, Config).

Change-Id: I36703bdf8d2fb893477fcca5f56b671ca8e3659e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>