[FAB-4102] Proto translator statically opaque comp
The proto translation framework introduced in FAB-4100 needs to be
applied to messages with opaque fields.  This is one of the key goals of
proto translation, which prevents the protos as defined from being human
readable.

This CR adds a statically opaque proto msg handler to the proto
translation framework.

Change-Id: Ifebc46f3bd3ec2087a9f2d579de30da5a3cb784d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>