[FAB-4103] Proto translator variably opaque comp
The proto translation framework introduced in FAB-4100 needs to be
applied to messages with opaque fields.  This is one of the key goals of
proto translation, which prevents the protos as defined from being human
readable.

This CR adds a variably opaque proto msg handler to the proto
translation framework.  Documented further in api.go, but variably
opaque fields differ from statically opaque ones, in that variably
oapque fields may depend upon the contents of the proto being populated
to determine type information at runtime and will be evaluated after the
static ones.

Change-Id: Iacbba1316f8e82dee68695aa98bf4c30a1a139da
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>