[FAB-4104] Proto translator dynamic field comp
The proto translation framework introduced in FAB-4100 needs to be
applied to messages with dynamic fields.  Dynamic fields are fields
whose message may need to be unmarshaled differently depending on
context.

For instance, different ConfigValue messages may resolve the same field
to different types, depending on the context in which the ConfigValue is
evaluated.

This CR adds a dynamic proto msg handler to the proto translation
framework.

Change-Id: I4bf809721f9af37778aeb3df3d706bfda1ed4e7b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>