[FAB-4085] Prevent expiration of self identity
In gossip we have time-based expiration (purge) of identities that haven't
been used for a long time.
This may cause an expiration of the peer's own certificate if it doesn't
gossip with any peers (i.e if it's alone in the world).

The problem is that when an identity expires from the identity store,
it is also expired from the pull mediator that is used for identities.
This is important because the only way identities are gossiped
transitively is via the pull mechanism.
If a peer's own identity disappears from the pull mediator,
it will never be sent to peers transitively.

Change-Id: Ic2138ebaa60bdf2454d65d3884f141c3736254a0
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>