[FAB-4106] Config update computation library
As part of FAB-1678, the configtx translator helper tool needs to be
able to take in an original config, a modified config, and compute the
config update transaction which would transition between the two.

This CR adds a small package for performing this computation, to be
leveraged by the actual tool code in FAB-1678.

Change-Id: I931d316e464dfb95589293d84ad499693ba11d4c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>