[FAB-4208] proper UT for chaincode framework
Replaces the heavyweight chaincode framework tests that
need a full fledged peer and container env for testing with
a lightweight UT tests using a mock environment. The CC side
and some of the ledger interfaces are mocked.

Also disables other tests in the framework so the community
can get a feel for the coverage and speedup

. renamed cc_test.go to chaincode_support_test.go
. add some minor tests
. added some error refactoring in handler.go

Change-Id: I462f05f7a78b7b2cfbb20138a20c5c0ee45a9be4
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>