[FAB-4205] Clarify missing system chain error
As reported in FAB-4205, it can be confusing for a user who generates a
genesis block via configtxgen without a consortiums group and attempts
to use this genesis block for orderer bootstrap.  The resulting error is
that no system channel is defined, and the orderer will not start.
However, it is non-obvious from the log what the underlying cause of the
error might be.

This CR adds a warning message to the configtxgen tool when it is being
used to generate a genesis block (typically used for orderer bootstrap)
which does not contain a consortiums group.  It additionally makes the
"No system channel" error message more verbose and suggests the likely
cause of a missing consortiums definition.

Change-Id: Ic3bb22e7ad401718d08de5f4d7a2918d2493dcc5
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>