[FAB-4302] Harden gossip bootstrap peer connection
Full details and introduction is in the JIRA item.

Gossip should be protect against mis-configurations that have security
implications, and therefore I suggest to make the bootstrap peer
connection to use the same mechanism as the anchor peer connection -
when a peer connects to an anchor peer, it first performs a handshake
(which is authenticated), and learns about the organization of the peer
behind the endpoint and only then decides whether to send its internal
endpoint or not.

Change-Id: Ieee1abde85f8e22c4ff6adad16dcd6fdab47b57f
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>