[FAB-4317] Fix go 1.7 uint32 parsing in protolator
Under go 1.7, the defaults for JSON unmarshaling are different than for
go 1.8.  In particular, go 1.7 unmarshals numbers to the float64 type by
default, while go 1.8 unmarshals numbers to the json.Numeric type (which
stores both an int64 and float64 representation).

The float64 numbers cannot be converted to uint32 fields in the proto
structures, causing an error in decoding.

This CR switches from using the default json.Unmarshal method to a
custom json.Decoder to achieve the go 1.8 behavior in go 1.7.

Change-Id: I1d3a980f116dddb1749dc9110ffb4a94f2cc105c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>