[FAB-4163] Nil dereference in configtxgen
When attempting to generate a genesis block while failing to specify an
orderer section, the result was a nil point dereference.

This CR fixes the dereference by protecting it behind a nil check, and
updates the tool logic to check for the existence of the section, and
print an error and exit if it is not present.

Change-Id: Id768f23d7e138fd7c8ca90889e805975fdf44bec
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>