[FAB-4330] Orderer panic on restart
On restart, the orderer is failing to properly re-initialize the config
sequence number, which causes it to inappropriately flag new non-config
blocks as config blocks.

On second restart, the incorrect last config index causes the orderer to
attempt to treat a normal transaction as a config transaction.  The
error checking recognizes this as a fatal error and panics.

This CR moves the initialization out of the "new channel" path, and into
the common path shared by both startup and restart.

It also enhances the log messages to make detection of this sort of
scenario easier.

Change-Id: I0d730eb5180f21ee6913bc8633ff540128bc6899
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>