[FAB-4337] Fix proto style in batchsize config msg
This is a proto message that slipped through the proto style
normalization.

It's a safe change from a go API perspective, and it is ABI compatible
with the existing protos, so no existing consumers should be affected.

This is important to fix however, because reconfiguration is finally
available, and these field names are finally user visible.  Fixing this
soon should have minimal impact, but may have more impact in the future
once more people are familiar with the config contents.

Change-Id: I7b21f72776088e66be45c72ed869a43b7153b0d5
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>