[FAB-4342] Fix dev style channel creation
When the orderer is configured without any consortium members, channels
may be created with no members.  This is an insecure mode, and supported
only for development or test environments.

However, sometimes it is desirable to create channels with consortium
members while in this insecure mode, but the current code rejects such
requests because the consortium member is not present.

This CR modifies the orderer channel creation code to allow the second
use case while in the insecure mode.

Change-Id: Ibf37ab1dbdf79b20e1238ad128fccebf04193dfd
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>