[FAB-4360] Add release-clean to clean target
When running the e2e_cli, if you have
previously built the release artifacts
you may encounter issues.

Given that it's normal to run
make clean prior to tests, etc,
adding release-clean to the
clean target is required.

Also noticed that the generateArtifacts
script was building releases for all
platforms so modified it to use the
target for the current platform it's
being run on as well.

Change-Id: Ic28ae812b9f9138fb69d75d87de4b1c74f3bcf4b
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>