[FAB-4352] Add version cmd to configtxlator
Adds a version cmd to configtxlator.
Modified configxtxlator to use kingpin
as the CLI library as that's what is
used in cryptogen as well.

Version info is added to the binary
using both
make configtxlator
make release[-all]

(Note: this fixes the issue with
make cryptogen not adding the version
as well)

Change-Id: I6bc52f12a09b4ea60dcab2693205cf151fb1b681
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>