[FAB-4418] Fix confusing policy naming
When traversing the JSON tree as generated by configtxlator, the
following list of elements is encountered:

policies -> <name> -> policy -> policy -> policy -> n_out_of -> policies

This is understandably very confusing for users, as so many different
elements claim to represent a policy.  This CR modifies the policy
definitions so that instead that same list of elements becomes:

policies -> <name> -> policy -> value -> rule -> n_out_of -> rules

This makes it much more clear that the third element is a named policy,
and that the policy has a value, consisting of a rule, which itself may
be composed of other rules.

The consumers of the existing name scheme are restricted to those users
of configtxlator or those who have manually attempted to build these
messages (I am unaware of any).  This is not an ABI breakage, although
because of the nature of the bug it is an API breakage.

Change-Id: Ia0d99f6e08ce43bc2563dc2caf9de9fd512af191
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>