[FAB-4424] Clarify configtx flag and error
The configtxgen tool has a flag "-asOrg" which takes an organization's
name, and performs a config update as that org (useful for updating
anchor peers only, today).

Apparently this flag can be interpretted as requiring the organization's
MSP ID, and not name.

This CR adds a clarifying "(by name)" to the parameter help, and
additionally changes the error message when the org name is not found,
to refer to the organization name, and not simply the org.

Change-Id: I7c3dfc215d7efbf731ce0fb2464b9bafbd0572cb
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>