[FAB-3831] Report empty application/orderer groups
The config interface currently returns the application/orderer config
groups as an interface definition.  Because of this, a simple nil check
will not generally determine whether the config is actually set, and
consequently attempting to access the config may cause a nil pointer
dereference (even when the nil check returns false).

This CR adds a bool to this interface, to allow the caller to determine
whether the interface has been populated.

Change-Id: I37e0ba25dcacffb273e5712f072d65786206b05b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>