[FAB-3556] Throw error for invalid keys at simulation
This CR makes changes in the simulation such that the simulator
raises an error if an invalid key is passed in. The validity of the
key is decided based on the configured db. For instance, couchdb allows
only valid utf-8 strings to be the key.

In the current code, this check is not made during simulation. This causes
to defer the issue at the commit time in which situation, peer has no option
other than raising a panic becasue, it simply cannot commit the block.

Change-Id: Id1439c36e3a142f8ca47574d391ee10b366acd69
Signed-off-by: manish <manish.sethi@gmail.com>