[FAB-4534] Use cancellation context in gossip
In gossip when a peer p closes a connection to another peer q
that it initiated connection with, it simply closes the gRPC stream,
but this doesn't always make the Recv() method to return and might
result in a stuck goroutine.

The solution would be to use a cancellation context when priming the stream object.

Change-Id: I6632333b09559b70670787418788674452749f0e
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>