[FAB-4590] Auto-vendor all deps properly
It was noted that the auto-vendoring mechanism introduced
in CR 6991 was failing to vendor dependencies properly.

Investigating, it was root-caused that we were only auto-
vendoring transitive dependencies, not direct imports.

This patch corrects this by merging the direct imports
into our dependency tree.  Fix was verified by adding
the unit-test, verifying that it fails, and subsequently
verifying that it passes after the fix was applied.

Credit to Rahul Hegde <rhegde@cls-bank.com> for finding the
issue and proposing the proper fix that inspired this patch.

Fixes FAB-4590

Change-Id: Ie0631ac1033c3f427ea42c43eccf76a0f7390320
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>