[FAB-4616] Skip type switch on unmarshaling error
Even though that path (error in unmarshaling + type switch) was
activated in the tests without issues, there is (might be? depends on
the `proto` package) a non-zero chance we can end up with a non-nil
error and a nil message. This will result in a nil panic dereference,
come type-switch time. A `continue` statement prevents this from
happening.

Change-Id: Iff99519a81e6cfafd0d118fa8eefa9a15fc42649
Signed-off-by: Kostas Christidis <kostas@christidis.io>