[FAB-4743] Harden backoff policy code in deliverClient
The delivery client has a backoff policy that specifies
an amount of time to sleep and whether to continue to retry
 based on an attempt number.

Currently the code that calculates the amount of time to sleep
is prone to overflow to a negative number.

However, the fact that the code only retries a limited number
of times (not indefinitely) saves us from the overflow.

I would like to make the code that calculates the time duration
to sleep to never overflow in order for if someone in the future
to change the attempt threshold and forget about the overflow problem
lurking around the corner, we won't have a problem.

Change-Id: Ie0d365e879eb71141c278231819e52e341b73fe5
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>