[FAB-4765] Remove dead MSP config code
The getMspConfig function loads the MSP signer certs from the signer
cert dir, error if the dir is not present, and then does nothing with
them.

This function is used to get the Verifying MSP definition to include in
a channel, so requiring that a signcerts directory be present is
non-sensical in this case, but still causes an error if not present.

Change-Id: Ic89709df51f768a06c08551563b1e00df05f7695
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>