[FAB-4830] cauthdsl signature verify err unlogged
We have seen in a number of bug reports, such as FAB-4827 that the
output provided by cauthdsl in the logs is cryptic, and hard to debug
from.  Additionally, the error returned by the signature verification is
not actually logged, which makes conclusive debugging nearly impossible.

This CR reworks the logging messages to be more uniform, less cryptic,
and logs the particular err which was ignored.

Change-Id: I229ee62be881d5d136a4532174990386d644853e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>