[FAB-4243] Set CouchDB max_dbs_open config option
CouchDB returns errors to peer if more than N channels (databases)
are being accessed concurrently.

Current default value for max_dbs_open in default.ini is 500.
This corresponds to 500 shards or 62 databases (there are 8 shards per
database by default).  For highly concurrent environments with a large
number of channels, a better default would be 1000 databases, which
corresponds to max_dbs_open of 8000.

CouchDB uses LRU cache to manage open databases,
and closes databases as needed.

Perhaps more important than picking a higher default value, is that
the configuration value be exposed in the CouchDB local.ini
configuration file, so that consumers are aware of the setting,
and can easily adjust if needed.

Change-Id: I19f99737939ed27c70955c6d5784232f02ef8b2c
Signed-off-by: David Enyeart <enyeart@us.ibm.com>