[FAB-4884] Restrict KeyUsage for signing keys
FAB-4626 introduces a change to restrict the
KeyUsage for signing certificates to
x509.KeyUsageDigitalSignature and restricts
ExtendedKeyUsage as well.

cryptogen currently uses the same certs
for signing and TLS and therefore the
the generated certs will no longer be
valid signing identities.

This CR does the following:
- generates signing certs with the proper
KeyUsage
- generates separate TLS certs with the
appropriate KeyUsage and ExtendedKeyUsage

Change-Id: I9112d47d2635370489fec3ff07f09118dc021318
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>