[FAB-4751] allow provisioning separate chaincode server
The chaincodeListenAddress property allows chaincode connections to use
a listener other than the peer's listener. With this, chaincode streams
need not be bound to the peers endpoint allowing more network options for
protecting it.

The e2e_cli sample uses the chaincodeListenAddress to excercise the
separate chaincode listener (on 7052 but on the same IP address as the
that peer's listener uses).

patch 1
. comment changes (thanks, Binh)

Change-Id: Ic0943be6739ef0401e12686721d3e0472479580c
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>