[FAB-4963] update byfn instructions for tls
Orderer now uses separate tls ca certs.
Updating byfn manual instructions to utilize
the new orderer tls ca certs. This is required
for the manual commands to work.

Also, there was a mix of <channel_id> and
$CHANNEL_NAME in the manual command instructions.
Made all commands consistent to use $CHANNEL_NAME
so that an environment variable can be set,
which enables simple copy/paste of all the commands.

Change-Id: Ia1b978bc192896eef10b1cac59a931f82d65c2a3
Signed-off-by: David Enyeart <enyeart@us.ibm.com>